Add files via upload
Project Predictive Analytics on "Retail Data" data is an increase and decrease in transactions every year and the most significant decrease occurred in 2019 where previously in 2018 the number of customers was still very high. Because of these problems, the results of data analysis obtained in this project can be used as a reference for business solutions for companies, especially to help store managers to innovate and be creative again so that customers keep coming back to the store. In addition, this project is also useful to help solve this problem by creating models and data visualizations that can later be used and specifically studied models provide accuracy scores, precision scores, and recall scores that can help predict whether this effort will be successful or not. not so that the store manager can find out and find solutions to the problems the company is experiencing.